pyproject.toml
[project]
name = "rag-hr-chatbot"
version = "0.1.0"
description = "Chatbot RAG para FAQs de HR SaaS"
authors = [
{ name = "CMelon" }
]
requires-python = ">=3.10"

dependencies = [
    # ── OpenAI SDK + LangChain integration ────────────────────────────────────
    "openai>=1.0.0",
    "langchain-openai>=0.1.0",

    # ── LangChain core ─────────────────────────────────────────────────────────
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",

    # ── LangGraph para agentes ─────────────────────────────────────────────────
    "langgraph>=0.2.0",

    # ── RAG: vector stores + embeddings ───────────────────────────────────────
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",

    # ── Utilidades ─────────────────────────────────────────────────────────────
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "pypdf>=4.0.0",
    "rich>=13.0.0"
]

[tool.setuptools]
packages = ["src"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
